Fix CMake install DIRECTORY
This commit modifies CMakeLists.txt to fix the launch directory
installation.  We want the entire launch directory to be installed as
a subdirectory.

We want:

 launch/*  -->  install/<package_name>/share/<package_name>/launch/*

To do this, we do:

install(DIRECTORY
  launch         # do not include the traiing slash, like "launch/"
  DESTINATION share/${PROJECT_NAME}
  )

If we use "launch/" instead of "launch", CMake will not create the
new sub-directory and we end up:

 launch/*  -->  install/<package_name>/share/<package_name>/*

We lost the "launch" sub-directory hierarchy in the install directory.